+++ src/GithubFlavoredMarkdownConverter.php
@@ -26,8 +26,14 @@ class GithubFlavoredMarkdownConverter extends CommonMarkConverter
     {
         if ($environment === null) {
             $environment = Environment::createGFMEnvironment();
+        } else {
+            @\trigger_error(\sprintf('Passing an $environment into the "%s" constructor is deprecated in 1.6 and will not be supported in 2.0; use MarkdownConverter instead', self::class));
         }
 
-        parent::__construct($config, $environment);
+        if ($environment instanceof ConfigurableEnvironmentInterface) {
+            $environment->mergeConfig($config);
+        }
+
+        MarkdownConverter::__construct($environment);
     }
 }